+++ packages/web-framework/bin/static/scripts/ti-framework.js
@@ -518,6 +518,26 @@ let configureApplication = () => {
         },
 
         /**
+         * Redirect the user to the specified screen.
+         *
+         * @method
+         * @param {string} screen
+         * @public
+         */
+        openScreen( screen ) {
+            if ( !screen || !/^[\w-]+$/.test( screen ) || !window.htmx ) {
+                window.location.href = "/";
+            } else {
+                let screenUrl = "/app/" + screen;
+                window.htmx.ajax( "get", screenUrl, { target: "#ti-content", swap: "innerHTML" } ).then( () => {
+                    window.history.pushState( null, "", screenUrl );
+                } ).catch( () => {
+                    window.location.href = "/";
+                } );
+            }
+        },
+
+        /**
          * Used to format an exception notification message.
          *
          * @method